requirements.txt
fastapi
uvicorn
streamlit
requests
sqlalchemy
pydantic[email]
python-multipart
python-dotenv

langchain
langchain-classic
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langchain-chroma
chromadb

FlagEmbedding
transformers
torch
